Upgrade to .NET 10 and SpotifyAPI.Web 7, load config from user secrets
Retarget the app from netcoreapp3.1 to net10.0 and modernize the
SpotifyAPI.Web integration from the legacy v5 SpotifyWebAPI God-object
to the immutable v7 ISpotifyClient API.

Highlights:
- csproj: net10.0, SpotifyAPI.Web 7.4.2, AspNet.Security.OAuth.Spotify
  10.0.0, EF Core Sqlite 10.0.9, Serilog stack, BouncyCastle.Cryptography,
  Humanizer, explicit Newtonsoft.Json. Added <UserSecretsId>.
- Spotify v5 to v7 migration across SpotifyHelpers, SpotifyAccessService,
  InternalController and PartyRoom. All calls use current, non-obsolete
  endpoints (Feb 2026 Spotify API changes accounted for): Playlists.
  CurrentUsers, Playlists.Create(request), GetPlaylistItems,
  RemovePlaylistItems (V2), Library.SaveItems, Player.* and
  UserProfile.Current.
- Merge Startup.cs into Program.cs using minimal hosting (WebApplication).
- Config now prefers .NET user secrets, then falls back to environment
  variables (AddUserSecrets added after env vars so secrets win).
- SignalR models: switch [JsonIgnore] to System.Text.Json so ConnectionId
  and TargetId are no longer serialized to clients.
