+++ sources/PugetSound/PugetSound/Logic/RoomMember.cs
@@ -1,4 +1,4 @@
-﻿using Newtonsoft.Json;
+﻿using System.Text.Json.Serialization;
 
 namespace PugetSound.Logic
 {
